+++ changelog.md
@@ -2,6 +2,8 @@
 
 **urlfix 0.2.1**
 
+* `dirurlfix` is a new module dedicate to directory replacements. 
+
 * Fixed issues with markdown links in the format `[]()[]()` not being fully matched. 
  See [#17](https://github.com/Nelson-Gon/urlfix/issues/17)